requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
langchain-google-genai
langchain-core
ollama
selenium
webdriver-manager
pypdf
pytesseract
Pillow
google-generativeai
requests
sqlalchemy
asyncpg
greenlet
